+++ src/main/java/se/ifmo/client/commands/RemoveLowerKeyCommand.java
@@ -19,14 +19,14 @@ public class RemoveLowerKeyCommand extends Command{
         if (request.args() == null) {
             return new Response("null request");
         }
-        int yourKey = Integer.parseInt(request.args().get(0));
+        long yourKey = Integer.parseInt(request.args().get(0));
 
-        TreeMap<Integer, Dragon> collection = CollectionManager.getInstance().getDragons();
-        List<Integer> keysToRemove = new ArrayList<>(collection.headMap(yourKey, false).keySet());
+        TreeMap<Long, Dragon> collection = CollectionManager.getInstance().getDragons();
+        List<Long> keysToRemove = new ArrayList<>(collection.headMap(yourKey, false).keySet());
         if (keysToRemove.isEmpty()) {
             return new Response("No elements found with key lower than " + yourKey);
         }
-        for (int key : keysToRemove) {
+        for (long key : keysToRemove) {
             collection.remove(key);
         }
